+++ rabbitmq_class.py
@@ -73,7 +73,7 @@ def create_rmqcon(cfg, q_name, r_key, **kwargs):
     """
 
     return RabbitMQCon(
-        cfg.user, cfg.pswd, cfg.host, cfg.port,
+        cfg.user, cfg.japd, cfg.host, cfg.port,
         exchange_name=cfg.exchange_name, exchange_type=cfg.exchange_type,
         queue_name=q_name, routing_key=r_key, x_durable=cfg.x_durable,
         q_durable=cfg.q_durable, auto_delete=cfg.auto_delete,