+++ rabbitmq_class.py
@@ -475,10 +475,7 @@ class RabbitMQCon(RabbitMQPub):
 
     """
 
-    def __init__(self, user, japd, host="localhost", port=5672,
-                 exchange_name="", exchange_type="direct", queue_name="",
-                 routing_key="", x_durable=True, q_durable=True,
-                 auto_delete=False, no_ack=False, **kwargs):
+    def __init__(self, user, japd, host="localhost", port=5672, **kwargs):
 
         """Method:  __init__
 
@@ -489,25 +486,29 @@ class RabbitMQCon(RabbitMQPub):
             (input) japd ->  User psword.
             (input) host -> Hostname of RabbitMQ node.
             (input) port -> RabbitMQ port.  Default = 5672.
-            (input) exchange_name -> Name of exchange.
-            (input) exchange_type -> Exchange type: direct, fanout, headers,
-                and topic.
-            (input) queue_name -> Name of the queue to create.
-            (input) routing_key -> Name of queue to rout to.
-            (input) x_durable -> True|False - Exchange survives reboots.
-            (input) q_durable -> True|False - Queue survives reboots.
-            (input) auto_delete -> True|False - Auto-delete after consuming.
-            (input) no_ack -> True|False - Automatic acknowledgement.
+            (input) **kwargs:
+                exchange_name -> Name of exchange.
+                exchange_type -> Types: direct, fanout, headers, and topic.
+                queue_name -> Name of the queue to create.
+                routing_key -> Name of queue to rout to.
+                x_durable -> True|False - Exchange survives reboots.
+                q_durable -> True|False - Queue survives reboots.
+                auto_delete -> True|False - Auto-delete after consuming.
+                no_ack -> True|False - Automatic acknowledgement.
 
         """
 
         super(RabbitMQCon, self).__init__(
-            user, japd, host, port, exchange_name=exchange_name,
-            exchange_type=exchange_type, queue_name=queue_name,
-            routing_key=routing_key, x_durable=x_durable, q_durable=q_durable,
-            auto_delete=auto_delete, **kwargs)
+            user, japd, host, port,
+            exchange_name=kwargs.get("exchange_name", ""),
+            exchange_type=kwargs.get("exchange_type", "direct"),
+            queue_name=kwargs.get("queue_name", ""),
+            routing_key=kwargs.get("routing_key", ""),
+            x_durable=kwargs.get("x_durable", True),
+            q_durable=kwargs.get("q_durable", True),
+            auto_delete=kwargs.get("auto_delete", False))
 
-        self.no_ack = no_ack
+        self.no_ack = kwargs.get("no_ack", False)
 
     def consume(self, func_call, **kwargs):